Fix critical bugs from code review + validate TT/killer moves against movelist
Critical fixes:
- EP capture direction inverted in NNUE accumulator update (was +8/-8, now -8/+8)
- Add repetition detection via undo stack hash comparison
- Fix i64->i32 overflow in NNUE forward pass (divide in i64 before cast)

Move validation:
- TT/killer/counter moves now verified against generated move list (from/to match)
- Geometric validation for piece reachability in is_pseudo_legal
- moves_tried counter now incremented after make_move succeeds (not before)

Performance:
- Stack-allocate NNUE change list (was Vec::with_capacity(5) per move)
- Stack-allocate quiets_tried array (was Vec per node)
- CLI -nnue flag for pre-loading NNUE net before UCI loop

+++ src/nnue.rs
@@ -404,7 +404,7 @@ impl NNUENet {
                 }
             }
 
-            let mut result = (output as i32) * EVAL_SCALE / QAB;
+            let mut result = (output * EVAL_SCALE as i64 / QAB as i64) as i32;
             if self.use_screlu {
                 result = result * 4 / 5;
             }
@@ -438,8 +438,8 @@ impl NNUENet {
         }
 
         // Scale: output is at QA × QB = 16320
-        // centipawns = output * 400 / 16320
-        let mut result = (output as i32) * EVAL_SCALE / QAB;
+        // centipawns = output * 400 / 16320 (do division in i64 to avoid overflow)
+        let mut result = (output * EVAL_SCALE as i64 / QAB as i64) as i32;
 
         // SCReLU scale correction: squared activation has wider dynamic range
         if self.use_screlu {
